backend_python/requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
Pillow==10.1.0
PyPDF2==3.0.1
pandas==2.1.3
python-dotenv==1.0.0
cloudinary==1.36.0
qdrant-client==1.6.9
google-generativeai==0.3.2
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-csv==0.0.13
uuid==1.30
pydantic==2.5.0
httpx==0.25.2
asyncio==3.4.3
aiohttp==3.9.1
